npm rule: avoid string allocation for parsed pkg name
This was remarked also in the original PR
which added the Npm rule
(https://github.com/AikidoSec/safechain-agent/pull/14)

but it was left unactioned. This PR shows what was meant
with that. I consider these important enough as parsing of
request uris is in the hot path of the proxy. As such it is
important to be mindful of the resources we consume
of the user, given the proxy is to run on end-user devices.